Small style fixes in B3MoveConstants.cpp https://bugs.webkit.org/show_bug.cgi?id=151980
Reviewed by Benjamin Poulain.

* b3/B3MoveConstants.cpp:


Canonical link: https://commits.webkit.org/170124@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@193687 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/JavaScriptCore/b3/B3MoveConstants.cpp
@@ -247,7 +247,7 @@ private:
                 origin);
             result = m_insertionSet.insert<MemoryValue>(
                 valueIndex, Load, key.type(), origin, tableBase,
-                sizeof(intptr_t) * m_constTable.get(key));
+                sizeof(int64_t) * m_constTable.get(key));
         } else
             result = m_insertionSet.insertValue(valueIndex, key.materialize(m_proc, origin));
         m_constants.add(key, result);
@@ -257,7 +257,7 @@ private:
     bool goesInTable(const ValueKey& key)
     {
         return (key.opcode() == ConstDouble && key != doubleZero())
-        || (key.opcode() == ConstFloat && key != floatZero());
+            || (key.opcode() == ConstFloat && key != floatZero());
     }
 
     bool needsMotion(const ValueKey& key)
